+++ js/main.js
@@ -1,6 +1,4 @@
-import './util.js';
-import './data.js';
+import { createPhotos } from './mock.js';
 import { generateRandomUsersPics } from './miniature.js';
-import { createPhotos } from './data.js';
 createPhotos();
 generateRandomUsersPics();
